refactor: QueryGenerateViewModel -> QueryGenreateManger로 변경
